♻️ refactor(tools): flatten core tools into internal/tools package
Merge coretools/, bash/, read/, write/, edit/, notify/ sub-packages
into a single `internal/tools` package (package coretools). Each tool
gets its own file. Complex tools (delegate, mcp, skills) remain as
sub-packages.

- Delete standalone bash/read/write/edit packages (were dead code)
- Promote coretools/ contents to internal/tools/
- Move notify tool into internal/tools/notify.go
- Add Definitions() for server API metadata without sandbox
- Directory name is `tools`, package name is `coretools`
