--- internal/tools/edit/edit_test.go
@@ -1,163 +1,0 @@
-package edit
-
-import (
-	"context"
-	"os"
-	"path/filepath"
-	"strings"
-	"testing"
-)
-
-func TestEditTool_Definition(t *testing.T) {
-	tool := NewEditTool("")
-	def := tool.Definition()
-	if def.Name != "edit" {
-		t.Errorf("expected name 'edit', got %q", def.Name)
-	}
-}
-
-func TestEditTool_BasicReplace(t *testing.T) {
-	dir := t.TempDir()
-	path := filepath.Join(dir, "file.txt")
-	if err := os.WriteFile(path, []byte("hello world"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	tool := NewEditTool("")
-	result, err := tool.Execute(context.Background(), map[string]any{
-		"path":       path,
-		"old_string": "world",
-		"new_string": "Go",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	data, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if string(data) != "hello Go" {
-		t.Errorf("expected 'hello Go', got %q", string(data))
-	}
-	if result == "" {
-		t.Error("expected non-empty result message")
-	}
-}
-
-func TestEditTool_ResolvesRelativePathFromProjectRoot(t *testing.T) {
-	projectRoot := t.TempDir()
-	path := filepath.Join(projectRoot, "nested", "file.txt")
-	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(path, []byte("hello world"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	tool := NewEditTool(projectRoot)
-	_, err := tool.Execute(context.Background(), map[string]any{
-		"path":       "nested/file.txt",
-		"old_string": "world",
-		"new_string": "Stella",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	data, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if string(data) != "hello Stella" {
-		t.Fatalf("expected edited content, got %q", string(data))
-	}
-}
-
-func TestEditTool_RejectsRelativePathWithoutProjectRoot(t *testing.T) {
-	tool := NewEditTool("")
-	_, err := tool.Execute(context.Background(), map[string]any{
-		"path":       "relative.txt",
-		"old_string": "x",
-		"new_string": "y",
-	})
-	if err == nil || err.Error() == "" || !strings.Contains(err.Error(), "requires a project root") {
-		t.Fatalf("expected project root error, got %v", err)
-	}
-}
-
-func TestEditTool_NotFound(t *testing.T) {
-	dir := t.TempDir()
-	path := filepath.Join(dir, "file.txt")
-	if err := os.WriteFile(path, []byte("hello world"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	tool := NewEditTool("")
-	_, err := tool.Execute(context.Background(), map[string]any{
-		"path":       path,
-		"old_string": "nonexistent",
-		"new_string": "x",
-	})
-	if err == nil {
-		t.Error("expected error when old_string not found")
-	}
-}
-
-func TestEditTool_AmbiguousMatch(t *testing.T) {
-	dir := t.TempDir()
-	path := filepath.Join(dir, "file.txt")
-	if err := os.WriteFile(path, []byte("foo foo"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	tool := NewEditTool("")
-	_, err := tool.Execute(context.Background(), map[string]any{
-		"path":       path,
-		"old_string": "foo",
-		"new_string": "bar",
-	})
-	if err == nil {
-		t.Error("expected error for ambiguous (multi-match) old_string")
-	}
-}
-
-func TestEditTool_MissingPath(t *testing.T) {
-	tool := NewEditTool("")
-	_, err := tool.Execute(context.Background(), map[string]any{
-		"old_string": "x",
-		"new_string": "y",
-	})
-	if err == nil {
-		t.Error("expected error when path is missing")
-	}
-}
-
-func TestEditTool_MissingOldText(t *testing.T) {
-	dir := t.TempDir()
-	path := filepath.Join(dir, "file.txt")
-	if err := os.WriteFile(path, []byte("content"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	tool := NewEditTool("")
-	_, err := tool.Execute(context.Background(), map[string]any{
-		"path":       path,
-		"new_string": "y",
-	})
-	if err == nil {
-		t.Error("expected error when old_string is missing")
-	}
-}
-
-func TestEditTool_FileNotExist(t *testing.T) {
-	tool := NewEditTool("")
-	_, err := tool.Execute(context.Background(), map[string]any{
-		"path":       "/nonexistent/path/file.txt",
-		"old_string": "x",
-		"new_string": "y",
-	})
-	if err == nil {
-		t.Error("expected error for non-existent file")
-	}
-}
