+++ internal/tools/normalize_test.go
