+++ internal/tools/notify.go
@@ -1,4 +1,4 @@
-package notify
+package coretools
 
 import (
 	"context"
@@ -7,24 +7,23 @@ import (
 	pkgchannel "github.com/CherryHQ/stella/pkg/channel"
 	"github.com/CherryHQ/stella/pkg/memory"
 	pkgplugins "github.com/CherryHQ/stella/pkg/plugins"
-	"github.com/CherryHQ/stella/pkg/tools"
+	pkgtools "github.com/CherryHQ/stella/pkg/tools"
 )
 
-// NewTool creates a notify tool with the given notifier service.
+// NewNotifyTool creates a notify tool with the given notifier service.
 // Returns nil if notifier is nil.
-func NewTool(notifier pkgplugins.Notifier) *Tool {
+func NewNotifyTool(notifier pkgplugins.Notifier) pkgtools.Tool {
 	if notifier == nil {
 		return nil
 	}
-	return &Tool{service: notifier}
+	return &notifyTool{service: notifier}
 }
 
-// Tool is an agent tool that sends notifications through the plugin host.
-type Tool struct {
+type notifyTool struct {
 	service pkgplugins.Notifier
 }
 
-var inputSchema = map[string]any{
+var notifyInputSchema = map[string]any{
 	"type": "object",
 	"properties": map[string]any{
 		"message": map[string]any{
@@ -47,15 +46,15 @@ var inputSchema = map[string]any{
 	"required": []string{"message"},
 }
 
-func (t *Tool) Definition() tools.Definition {
-	return tools.Definition{
+func (t *notifyTool) Definition() pkgtools.Definition {
+	return pkgtools.Definition{
 		Name:        "notify",
 		Description: "Send a notification message to the user. In normal user conversations, omit 'chat_id' so Stella can route via the current user's linked identities automatically. Supports multiple backends (Telegram, Slack, etc.). Use this for proactive messages, alerts, scheduler summaries, or long-running task results.",
-		InputSchema: inputSchema,
+		InputSchema: notifyInputSchema,
 	}
 }
 
-func (t *Tool) Execute(ctx context.Context, args map[string]any) (string, error) {
+func (t *notifyTool) Execute(ctx context.Context, args map[string]any) (string, error) {
 	message, _ := args["message"].(string)
 	if message == "" {
 		return "", fmt.Errorf("message is required")
