+++ internal/tools/tools.go
@@ -1,0 +1,67 @@
+package coretools
+
+import (
+	"github.com/CherryHQ/stella/pkg/sandbox"
+	pkgtools "github.com/CherryHQ/stella/pkg/tools"
+)
+
+// New returns the four standard sandbox-backed tools (bash, read, write, edit).
+func New(host sandbox.Host, toolsBinDir, projectRoot string) []pkgtools.Tool {
+	if host == nil {
+		return nil
+	}
+	return []pkgtools.Tool{
+		newBashTool(host, toolsBinDir, projectRoot),
+		newReadTool(host, projectRoot),
+		newWriteTool(host, projectRoot),
+		newEditTool(host, projectRoot),
+	}
+}
+
+// Definitions returns the canonical definitions for all core tools.
+// No sandbox session is required — useful for API metadata endpoints.
+func Definitions() []pkgtools.Definition {
+	return []pkgtools.Definition{
+		bashDefinition(),
+		readDefinition(),
+		writeDefinition(),
+		editDefinition(),
+	}
+}
+
+func resolveToolPath(host sandbox.Host, projectRoot, path string) (string, error) {
+	if projectRoot != "" {
+		return pkgtools.ResolveProjectPath(projectRoot, path)
+	}
+	return host.ResolvePath(path)
+}
+
+func resolveWritableToolPath(host sandbox.Host, projectRoot, path string) (string, error) {
+	if projectRoot != "" {
+		resolved, err := pkgtools.ResolveProjectPath(projectRoot, path)
+		if err != nil {
+			return "", err
+		}
+		return host.ResolveWritePath(resolved)
+	}
+	return host.ResolveWritePath(path)
+}
+
+func toolIntArg(args map[string]any, key string, defaultVal int) int {
+	v, ok := args[key]
+	if !ok {
+		return defaultVal
+	}
+	switch n := v.(type) {
+	case float64:
+		return int(n)
+	case int:
+		return n
+	case int32:
+		return int(n)
+	case int64:
+		return int(n)
+	default:
+		return defaultVal
+	}
+}
